Rework codebase, redefine models and add alembic migration
Codebase Rework
1. Re-initialise the project using poetry
2. Change the directory structure
3. Remove postgres database data-directory
3. Change from relative to absolute imports

Redefine Models
1. Add SQLAlchemy mixin to automatically generate a custom 8-digit UUID string
2. Change column names to be used in the models

Add initial alembic migration
1. Auto Generate initial migration file using alembic
2. Create database tables using the initial revision file
